refactor to extract WebSocketClient interface
- don't require javax.websocket based client/server impls
- ws based impl is now Jsr356WebSocketClient in package
...wsclient.javax.websocket  ugh?
