+++ connectors/common/build.xml
@@ -5,17 +5,18 @@
         Build common connector code.
     </description>
 
+  <property name="component.path" value="connectors/common"/>
   <import file="../../common-build.xml"/>
 
   <path id="compile.classpath">
-    <pathelement location="${lib}/quarks.api.topology.jar" />
+    <pathelement location="${quarks.lib}/quarks.api.topology.jar" />
     <path refid="quarks.ext.classpath" />
   </path>
 
 
   <path id="test.compile.classpath">
     <pathelement location="${jar}" />
-    <pathelement location="${lib}/quarks.providers.direct.jar"/>
+    <pathelement location="${quarks.lib}/quarks.providers.direct.jar"/>
     <pathelement location="../../api/topology/test.classes"/>
     <pathelement location="../../providers/direct/test.classes"/>
     <path refid="compile.classpath"/>
