+++ connectors/file/build.xml
@@ -10,7 +10,6 @@
 
   <path id="compile.classpath">
     <pathelement location="${quarks.lib}/quarks.api.topology.jar" />
-    <pathelement location="${quarks.lib}/quarks.connectors.common.jar" />
     <path refid="quarks.ext.classpath" />
   </path>
