+++ connectors/jdbc/build.xml
@@ -11,7 +11,6 @@
 
   <path id="compile.classpath">
     <pathelement location="${quarks.lib}/quarks.api.topology.jar" />
-    <pathelement location="${quarks.lib}/quarks.connectors.common.jar" />
   </path>
