+++ connectors/wsclient-javax.websocket/src/main/java/quarks/connectors/wsclient/javax/websocket/package-info.java
@@ -1,0 +1,8 @@
+/*
+# Licensed Materials - Property of IBM
+# Copyright IBM Corp. 2015,2016 
+*/
+/**
+ * WebSocket Client Connector for sending and receiving messages to a WebSocket Server.
+ */
+package quarks.connectors.wsclient.javax.websocket;
