+++ connectors/wsclient/src/main/java/quarks/connectors/wsclient/package-info.java
@@ -3,6 +3,6 @@
 # Copyright IBM Corp. 2015,2016 
 */
 /**
- * WebSocket Client Connector for sending and receiving messages to a WebSocket Server.
+ * WebSocket Client Connector API for sending and receiving messages to a WebSocket Server.
  */
 package quarks.connectors.wsclient;
